% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Interval for the Mean}
\usage{
myci(x)
}
\arguments{
\item{x}{A numeric vector representing the sample data.}
}
\value{
A numeric vector of length 2 containing the lower and upper bounds of the 95% confidence interval for the mean.
}
\description{
Confidence Interval for the Mean
}
\examples{
x <- rnorm(30, mean = 10, sd = 12)
myci(x)
}
