% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydim.R
\name{mydim}
\alias{mydim}
\title{Dimesnsion Function}
\usage{
mydim(data)
}
\arguments{
\item{data}{an R object for example a matrix, array, or dataframe}
}
\value{
A numeric vector of length two. The first element is the number of rows,
and the second element is the number of columns in the dataset.
}
\description{
Dimesnsion Function
}
\examples{
## mydim('mtbe')   Returns the dimensions of the 'mtbe' dataset
}
