% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Distribution Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Numeric. The mean of the normal distribution.}

\item{sigma}{Numeric. The standard deviation of the normal distribution.}

\item{a}{Numeric. The x-value up to which the area under the curve will be shaded.}
}
\value{
A list of mean, standard deviation, and probability.
}
\description{
This function plots a normal distribution curve with mean `mu` and standard deviation `sigma`,
and shades the area under the curve from negative infinity to a specified value `a`.
It also calculates and returns the cumulative probability `P(X <= a)`.
}
\examples{
myncurve(2, 4, 3)
}
