% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Plot T Distribution and Calculate P-value}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{The t-value for which the p-value is to be calculated. This represents the observed test statistic.}

\item{xmax}{The maximum value on the x-axis for plotting the t-distribution. Default is 4.}

\item{n}{The sample size used to determine the degrees of freedom for the t-distribution. Default is 20.}

\item{alpha}{The significance level for hypothesis testing. Default is 0.05.}
}
\value{
A list containing: The critical t-value(s) at the specified alpha level and the computed two-tailed p-value.
}
\description{
Plot T Distribution and Calculate P-value
}
\examples{
mypvalue(t0 = 2.1)
mypvalue(t0 = 1.5, xmax = 5, n = 30, alpha = 0.01)
}
